refactor: Move static utils out of interface to helper class
